# feat: implement Issue #60 - Add stepCount parameter for plan validation