--- a/debug.js
+++ b/debug.js
@@ -1,0 +1,1 @@
+console.log("Debug: metadataPath exists:", await pathExists(metadataPath));