--- a/debug.js
+++ b/debug.js
@@ -1,1 +1,0 @@
-console.log("Debug: metadataPath exists:", await pathExists(metadataPath));