--- a/verification_output.txt
+++ b/verification_output.txt
@@ -1,0 +1,6 @@
+[0;34mIssue #59 Agent Validation Quick Verification Script[0m
+Description: Quick verification without npm commands
+Date: Mon Sep 15 21:46:08 CDT 2025
+
+[0;34m[CHECK][0m Checking if Create task tool exists
+[0;32m[PASS][0m Create task tool exists: src/tools/create-task.ts