… TEST-ERROR-PATTERNS.md
… src/core/TaskSplitDecisionEngine.ts
… src/logging/EventLogger.ts
… src/tools/archive-tasks.ts
… src/tools/check-tasks.ts
… src/tools/create-task.ts
… src/tools/get-full-lifecycle.ts
… src/tools/get-server-info.ts
… src/tools/get-task-context.ts
… src/tools/list-agents.ts
… src/tools/mark-complete.ts
… src/tools/read-task.ts
… src/tools/report-progress.ts
… src/tools/restore-tasks.ts
… src/tools/submit-plan.ts
… src/tools/sync-todo-checkboxes.ts
… src/tools/write-task.ts
… src/types.ts
… src/types/context-types.ts
… src/types/plan-metadata.ts
… src/utils/fs-extra-safe.ts
… src/utils/plan-parser.ts
… src/utils/validation.ts
… tests/integration/todowrite-sync.test.ts
… tests/unit/core/ContextThresholdDetector.test.ts
… tests/unit/core/HandoffContextGenerator.test.ts
… tests/unit/core/TaskSplitDecisionEngine.test.ts
… tests/unit/core/delegation-templates.test.ts
… tests/unit/core/guidance-templates.test.ts
… tests/unit/features/task-id-parameter.test.ts